requirements.txt
# Database
psycopg2-binary>=2.9.9
redis>=5.0.1

# Async
aiohttp>=3.9.1
asyncio>=3.4.3
asyncpg>=0.29.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-benchmark>=4.0.0

# Rate Limiting and Retries
ratelimit>=2.2.1
backoff>=2.2.1

# Type Hints
typing-extensions>=4.9.0

# Utilities
python-dotenv>=1.0.0  # For environment variables